# api-documentation-old.yaml
openapi: 3.0.0
info:
  title: InCrowd API
  version: 1.0.0
  description: |
    # InCrowd API - Documentazione Completa
    
    Benvenuto nella documentazione dell'API di InCrowd, una piattaforma innovativa di crowdsourcing per proposte e iniziative cittadine.
    
    ## Panoramica
    
    L'API InCrowd fornisce accesso completo alla piattaforma per:
    - Gestione utenti (cittadini privati ed enti)
    - Creazione e gestione proposte
    - Sistema di follow/followers
    - Amministrazione e moderazione
    
    ## Autenticazione e Sicurezza
    
    L'API utilizza autenticazione JWT (JSON Web Token):
    
    1. **Login**: Ottieni il token tramite `POST /api/auth/login`
    2. **Utilizzo**: Includi il token nell'header di ogni richiesta:
       ```
       Authorization: Bearer <your-jwt-token>
       ```
    3. **Scadenza**: I token hanno durata limitata, effettua refresh quando necessario
    
    ## Struttura delle Response
    
    Tutte le risposte seguono il formato standardizzato:
    
    ```json
    {
      "data": { /* Dati della risposta */ },
      "message": "Messaggio descrittivo",
      "error": null // Solo presente in caso di errori
    }
    ```
    
    ## Gestione Errori
    
    - **400 Bad Request**: Dati non validi o parametri mancanti
    - **401 Unauthorized**: Token mancante, scaduto o non valido
    - **403 Forbidden**: Permessi insufficienti per l'operazione
    - **404 Not Found**: Risorsa non trovata
...
    
    ## Tipi di Utente
    
...
    
...
    
...

...
servers:
...
paths:
...
  /api/follow/{userId}:
    post:
      tags: [Follow]
      operationId: followUser
      summary: Segui un utente
      description: |
        Inizia a seguire un utente specifico.
        Aggiorna i contatori di follower/following per entrambi gli utenti.
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            example: "60d0fe4f5311236168a109ca"
          description: "ID dell'utente da seguire"
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Follow creato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                data:
                  followId: "60d0fe4f5311236168a109cc"
                  followed:
                    _id: "60d0fe4f5311236168a109ca"
                    nome: "Mario"
                    cognome: "Rossi"
                    user_type: "privato"
                    followersCount: 151
                message: "User followed successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Follow già esistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                message: "Already following this user"
                error:
                  code: "ALREADY_FOLLOWING"
                  details: "You are already following this user"
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Follow]
      operationId: unfollowUser
      summary: Smetti di seguire un utente
      description: |
        Smette di seguire un utente specifico.
        Aggiorna i contatori di follower/following per entrambi gli utenti.
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            example: "60d0fe4f5311236168a109ca"
          description: "ID dell'utente da smettere di seguire"
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Unfollow effettuato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                data:
                  unfollowed:
                    _id: "60d0fe4f5311236168a109ca"
                    nome: "Mario"
                    cognome: "Rossi"
                    user_type: "privato"
                    followersCount: 149
                message: "User unfollowed successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Follow non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                message: "Follow relationship not found"
                error:
                  code: "FOLLOW_NOT_FOUND"
                  details: "You are not following this user"
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/follow/status/{userId}:
    get:
      tags: [Follow]
      operationId: getFollowStatus
      summary: Verifica stato follow
      description: |
        Verifica se l'utente corrente segue un utente specifico.
        Restituisce anche statistiche di follow per l'utente target.
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            example: "60d0fe4f5311236168a109ca"
          description: "ID dell'utente da verificare"
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Stato follow recuperato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                data:
                  isFollowing: true
                  isFollowedBack: false
                  user:
                    _id: "60d0fe4f5311236168a109ca"
                    nome: "Mario"
                    cognome: "Rossi"
                    user_type: "privato"
                    followersCount: 150
                    followingCount: 75
                message: "Follow status retrieved successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # =============================================================================
  # PROPOSTE ENDPOINTS
  # =============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token ottenuto tramite login"
  responses:
    BadRequest:
      description: "Richiesta non valida - dati mancanti o formato errato"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Invalid request data"
            error:
              code: "BAD_REQUEST"
              details: "The provided data is not valid"
    Unauthorized:
      description: "Non autorizzato - token mancante o non valido"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Authentication required"
            error:
              code: "UNAUTHORIZED"
              details: "Token is missing or invalid"
...
    NotFound:
      description: "Risorsa non trovata"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Resource not found"
            error:
              code: "NOT_FOUND"
              details: "The requested resource does not exist"
...
    InternalError:
      description: "Errore interno del server"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardResponse'
          example:
            message: "Internal server error"
            error:
              code: "INTERNAL_SERVER_ERROR"
              details: "An unexpected error occurred"
  schemas:
...
    StandardResponse:
      type: object
      properties:
        data:
          type: object
          description: "Contenuto della risposta (opzionale in caso di errore)"
        message:
          type: string
          description: "Messaggio informativo sempre presente"
          example: "Operation completed successfully"
        error:
          type: object
          description: "Dettagli errore (solo in caso di errore)"
          properties:
            code:
              type: string
              example: "ERROR_CODE"
            details:
              type: string
              example: "Detailed error description"
      required:
        - message

    # =============================================================================
    # USER SCHEMA (UNIFICATO)
    # =============================================================================
...
security:
  - BearerAuth: []
  - {}